# carehub-backend/docs/openapi/form-builder-v1.yaml
openapi: 3.1.0
info:
  title: CareHub Form Builder API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    API v1 for creating, assigning, completing, and scoring versioned forms. Builder/import/assignment
    administration requires `ADMIN`; assigned-form and draft/submit operations require `MANAGER`.
    Published versions are immutable. Creating a draft without `sections` clones the current published version.
    Breaking changes will be introduced under `/api/v2`; v1 will receive deprecation and sunset headers before retirement.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /forms/{formId}:
    parameters:
      - $ref: '#/components/parameters/FormId'
    get:
      tags: [Form Builder]
      operationId: getForm
      summary: Get form metadata
      responses:
        '200':
          description: Form found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Form Builder]
      operationId: updateForm
      summary: Update form metadata
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateFormRequest' }
      responses:
        '200':
          description: Form updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Form Builder]
      operationId: retireForm
      summary: Soft-delete and retire a form
      responses:
        '204': { description: Form retired }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /form-assignments/{assignmentId}:
    parameters:
      - name: assignmentId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Form Assignment]
      operationId: getFormAssignment
      summary: Get assignment detail (ADMIN)
      responses:
        '200':
          description: Assignment found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormAssignmentEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Form Assignment]
      operationId: revokeFormAssignment
      summary: Revoke an assignment and all active items (ADMIN)
      responses:
        '204': { description: Assignment revoked }
        '404': { $ref: '#/components/responses/NotFound' }
  /form-assignment-items/{assignmentItemId}:
    delete:
      tags: [Form Assignment]
      operationId: revokeFormAssignmentItem
      summary: Revoke one assigned form (ADMIN)
      parameters:
        - $ref: '#/components/parameters/AssignmentItemId'
      responses:
        '204': { description: Assignment item revoked }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /form-submissions/{submissionId}:
    parameters:
      - name: submissionId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Form Submission]
      operationId: getFormSubmission
      summary: Get submission detail
      responses:
        '200':
          description: Submission found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormSubmissionEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Form Submission]
      operationId: replaceFormSubmissionAnswers
      summary: Replace all answers in an owned draft
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateFormSubmissionRequest' }
      responses:
        '200':
          description: Draft updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormSubmissionEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Form Submission]
      operationId: cancelFormSubmissionDraft
      summary: Void an owned draft
      responses:
        '204': { description: Draft voided }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /form-submissions/{submissionId}/submission:
    post:
      tags: [Form Submission]
      operationId: submitFormSubmission
      summary: Validate, score, and make an owned draft immutable
      parameters:
        - name: submissionId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lockVersion]
              properties:
                lockVersion: { type: integer, format: int64, minimum: 0 }
      responses:
        '200':
          description: Submission completed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormSubmissionEnvelope' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    FormId:
      name: formId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    AssignmentItemId:
      name: assignmentItemId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
    FormStatus:
      type: string
      enum: [DRAFT, PUBLISHED, RETIRED]
...
    SubjectType:
      type: string
      enum: [USER, PATIENT, PROCESS, ROOM, DEPARTMENT]
...
    FormAssignmentStatus:
      type: string
      enum: [ACTIVE, REVOKED, EXPIRED]
...
    FormAssignment:
      type: object
      required: [id, manager, assignedBy, assignedAt, status, items]
      properties:
        id: { type: integer, format: int64 }
        manager: { $ref: '#/components/schemas/AssignmentUser' }
        assignedBy: { $ref: '#/components/schemas/AssignmentUser' }
        assignedAt: { type: string, format: date-time }
        validFrom: { type: [string, 'null'], format: date-time }
        validUntil: { type: [string, 'null'], format: date-time }
        revokedAt: { type: [string, 'null'], format: date-time }
        status: { $ref: '#/components/schemas/FormAssignmentStatus' }
        items:
          type: array
          items: { $ref: '#/components/schemas/FormAssignmentItem' }
    AssignmentUser:
      type: object
      required: [id, employeeCode, fullName]
      properties:
        id: { type: integer, format: int64 }
        employeeCode: { type: string }
        fullName: { type: string }
    FormAssignmentItem:
      type: object
      required: [assignmentItemId, formId, formCode, title, formVersionId, versionNumber, status]
      properties:
        assignmentItemId: { type: integer, format: int64 }
        formId: { type: integer, format: int64 }
        formCode: { type: string }
        title: { type: string }
        formVersionId: { type: integer, format: int64 }
        versionNumber: { type: integer }
        status: { $ref: '#/components/schemas/FormAssignmentStatus' }
...
    FormSubmissionStatus:
      type: string
      enum: [DRAFT, SUBMITTED, VOIDED]
    FormScoringStatus:
      type: string
      enum: [NOT_CONFIGURED, CALCULATED]
    FormSubmissionResult:
      type: [string, 'null']
      enum: [PASSED, FAILED_SCORE, FAILED_CRITICAL, null]
...
    UpdateFormSubmissionRequest:
      type: object
      required: [lockVersion, answers]
      properties:
        lockVersion: { type: integer, format: int64, minimum: 0 }
        answers:
          type: array
          items: { $ref: '#/components/schemas/FormAnswerRequest' }
    FormAnswerRequest:
      type: object
      required: [questionKey]
      properties:
        questionKey: { type: string, format: uuid }
        optionKey: { type: string, format: uuid }
        optionKeys:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
        textValue: { type: string, maxLength: 10000 }
        numberValue: { type: number }
        dateValue: { type: string, format: date }
        timeValue: { type: string, format: time }
    FormSubmission:
      type: object
      required: [id, assignmentItemId, formId, formCode, formVersionId, versionNumber, title, status, subject, scoringStatus, criticalFailure, answers, lockVersion]
      properties:
        id: { type: integer, format: int64 }
        assignmentItemId: { type: integer, format: int64 }
        formId: { type: integer, format: int64 }
        formCode: { type: string }
        formVersionId: { type: integer, format: int64 }
        versionNumber: { type: integer }
        title: { type: string }
        status: { $ref: '#/components/schemas/FormSubmissionStatus' }
        subject:
          type: object
          required: [type, employeeCode, fullName]
          properties:
            type: { type: string, const: USER }
            employeeCode: { type: string }
            fullName: { type: string }
            position: { type: [string, 'null'] }
            department: { type: [string, 'null'] }
        scoringStatus: { $ref: '#/components/schemas/FormScoringStatus' }
        result: { $ref: '#/components/schemas/FormSubmissionResult' }
        totalScore: { type: [number, 'null'] }
        maxScore: { type: [number, 'null'] }
        passingScore: { type: [number, 'null'] }
        convertedScore: { type: [number, 'null'] }
        criticalFailure: { type: boolean }
        scoreBreakdown:
          type: array
          items: { $ref: '#/components/schemas/FormScoreBreakdown' }
        answers:
          type: array
          items:
            type: object
            required: [questionKey, value]
            properties:
              questionKey: { type: string, format: uuid }
              optionKey: { type: [string, 'null'], format: uuid }
              value: { type: object, additionalProperties: true }
        lockVersion: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: [string, 'null'], format: date-time }
        submittedAt: { type: [string, 'null'], format: date-time }
    FormScoreBreakdown:
      type: object
      required: [questionKey, code, title, critical, baseScore, weight, weightedScore, maxScore]
      properties:
        questionKey: { type: string, format: uuid }
        code: { type: string }
        title: { type: string }
        critical: { type: boolean }
        baseScore: { type: number }
        weight: { type: number }
        weightedScore: { type: number }
        maxScore: { type: number }
    FormAssignmentEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/FormAssignment' }
...
    FormSubmissionEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/FormSubmission' }
...
    CreateFormRequest:
      type: object
      required: [code, title, subjectType]
      properties:
        code: { type: string, minLength: 2, maxLength: 50, pattern: '^\s*[A-Za-z0-9_-]+\s*$', example: SAFETY-ROUND }
        title: { type: string, minLength: 1, maxLength: 255 }
        description: { type: string, maxLength: 4000 }
        subjectType: { $ref: '#/components/schemas/SubjectType' }
        ownerDepartmentId: { type: integer, format: int64 }
    UpdateFormRequest:
      type: object
      required: [title, subjectType]
      properties:
        title: { type: string, minLength: 1, maxLength: 255 }
        description: { type: string, maxLength: 4000 }
        subjectType: { $ref: '#/components/schemas/SubjectType' }
        ownerDepartmentId: { type: integer, format: int64 }
...
    Form:
      allOf:
        - $ref: '#/components/schemas/CreateFormRequest'
        - type: object
          required: [id, status, createdAt]
          properties:
            id: { type: integer, format: int64, readOnly: true }
            status: { $ref: '#/components/schemas/FormStatus' }
            ownerDepartment: { type: object, additionalProperties: true }
            currentPublishedVersion: { type: object, additionalProperties: true }
            createdAt: { type: string, format: date-time, readOnly: true }
            updatedAt: { type: string, format: date-time, readOnly: true }
...
    FormEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/Form' }
...
    Error:
      type: object
      required: [error_code, message, correlation_id]
      properties:
        error_code: { type: string, example: VAL_001 }
        message: { type: string, example: Validation failed }
        correlation_id: { type: string, format: uuid }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: ADMIN role required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Form, version, or referenced department not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Duplicate code, existing draft, stale lock version, or immutable published version
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request or form schema validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...